Use the passed redirect_to value for checks in bp_core_login_redirect() instead of REQUEST global
The redirect_to value passed to the login_redirect filter in wp-login.php is a
cleaned up version of _REQUEST['redirect_to'], and is a more reliable indicator
of where the link is trying to send the user. So it's a good idea to use it
when checking for the presence of 'wp-admin' rather than referencing the
global directly.

See #4199

git-svn-id: http://buddypress.svn.wordpress.org/trunk@6034 cdf35c40-ae34-48e0-9cc9-0c9da1808c22