Strip shortcodes from comparison text before adding [Read More] link in activity stream
We test to see whether a 'Read More' link should be added to activity items
in the stream by comparing the excerpted version against the unprocessed text.
Since the excerpt is created using bp_create_excerpt() with the
'filter_shortcodes' flag set to true, we must also strip shortcodes on the
unprocessed text to get an accurate comparison for length.

This prevents the [Read More] link from being appended erroneously when
shortcodes are present in activity updates.

Fixes #4051. Props magnus78

git-svn-id: http://buddypress.svn.wordpress.org/trunk@6095 cdf35c40-ae34-48e0-9cc9-0c9da1808c22