Improves current_user_can( 'bp_moderate' ) implementation for 1.6 release
The newly introduced bp_moderate cap must be manually added to the
Administrator role on order for it to work properly under the current
implementation of BP caps, wherein they are not stored in the database like
other caps.

This is a workaround, which will be reverted for BP 1.7.

Allows other caps passing through bp_current_user_can() to be evaluated
properly, without being overridden for super admins.

Fixes #4296

git-svn-id: http://buddypress.svn.wordpress.org/trunk@6185 cdf35c40-ae34-48e0-9cc9-0c9da1808c22