Improved documentation in member-header.php
In member-header.php template, we suggest to theme authors that they can
include arbitrary profile data by calling a template function. Previously
we recommended bp_profile_field_data(). But this function, when called on a
field that contains values stored in an array, does not make the array
presentable. The newer bp_member_profile_data() is better for this purpose.
This changeset updates the documentation accordingly.

Fixes #4319

Props hnla

git-svn-id: http://buddypress.svn.wordpress.org/trunk@6218 cdf35c40-ae34-48e0-9cc9-0c9da1808c22