Use event delegation for removing group invites in bp-default
When jQuery.live() was replaced with jQuery.on() in r6635 #4714, it broke
some places in BP's JS where events must be bound to items added to the page
after document.ready(). In these cases, event delegation should be used. This
changeset fixes a problem where AJAX was broken for "Remove Invite" buttons on
group invitation pages for newly-added items on bp-default.

The same bug was fixed for bp-legacy in r6791. See #4763.

Fixes #4943

git-svn-id: http://buddypress.svn.wordpress.org/trunk@6922 cdf35c40-ae34-48e0-9cc9-0c9da1808c22