In BP_Button, make 'block_self' work in member loops
block_self is designed to make buttons invisible when a user is looking at his
own page. This changeset makes it so that block_self also hides block_self
buttons during a bp_has_members() loop on the iteration where the current user
id is the currently iterated users.

Note that BP does not currently put any such buttons in the members directory.
This patch is for plugins/themes that add their own custom buttons.

Fixes #4709

git-svn-id: http://buddypress.svn.wordpress.org/trunk@6924 cdf35c40-ae34-48e0-9cc9-0c9da1808c22