Improved UX for "Notify group members of changes" toggle
The "Notify group members...." toggle on the 'edit-details' screen of the group
admin is intended to allow the group administrator to send an email to the
group members when the group's name or description has changed. However, the
legacy UX - a radio button (suggesting a persistent setting), and the text
(which suggests persistent notifications) - have historically been misleading.
By changing the radio button to a checkbox, and making a slight modification to
the wording of the text, the purpose of the checkbox should be much clearer.

See #4344

git-svn-id: http://buddypress.svn.wordpress.org/trunk@6955 cdf35c40-ae34-48e0-9cc9-0c9da1808c22