Update 'Show all x comments' count after creating or deleting activity comments
On activity permalink pages, the comment list is collapsed if the item has more
than 6 responses, and a link is appended that reads 'Show all x comments'. This
changeset introduces additional JavaScript to update this count when a new
activity comment is created, or an existing one is deleted, via AJAX.

In the process of making the change, the logic for constructing the 'Show all
x comments' string was changed, so that it's now properly translatable.

Also cached a JS variable that was being used many times, to avoid unnecessary
DOM traversal.

Fixes #4844

git-svn-id: http://buddypress.svn.wordpress.org/trunk@7002 cdf35c40-ae34-48e0-9cc9-0c9da1808c22