Remove all users at the beginning of each testcase round
There's an oddity when running the test suite on multisite that causes user
1 not to be rolled back properly between tests. This messes up the unit test
sequence generator for usernames, because attempts to create new users through
the factory use duplicate user_login values. By wiping the users table at the
beginning of each test set, we ensure we're starting with a clean slate.

This is not an ideal solution - it may be a bug in the WP test suite - but it's
a harmless enough hack to get our own tests working properly for now.

git-svn-id: http://buddypress.svn.wordpress.org/trunk@7043 cdf35c40-ae34-48e0-9cc9-0c9da1808c22