Enforce DISTINCT in BP_Activity_Activity::get() queries
Previously, we've never really needed to state DISTINCT explicitly in
BP_Activity_Activity::get(), because the kinds of queries we ran would
naturally result in the exclusion of duplicates. The introduction of
the meta_query parameter in 1.8 - and in particular, the ability to
introduce more than one meta_query clause separated by the inclusive
OR relation - uncovered the bug, as single activity items were matching
more than one clause in the meta_query and thus being returned in the
results more than once.

Fixes #5118

git-svn-id: http://buddypress.svn.wordpress.org/trunk@7318 cdf35c40-ae34-48e0-9cc9-0c9da1808c22