Blogs: Modify recorded post permalink to use shortlink.
Previously, when BuddyPress recorded the post permalink in the
activity stream, it would use the current permalink structure of the blog.

The problem with this is if the blog changed its permalink structure,
the post permalink in the activity post would not resolve properly
because of the change.

This commit alters the recorded post permalink to use the shortlink
(domain/?p=POST_ID).  Because WP's rewrite rules will pick up the
'p' query parameter and redirect to the proper post permalink, this
will ensure that the post permalink for future activity blog posts
will work as intended.

Fixes #4367.

git-svn-id: http://buddypress.svn.wordpress.org/trunk@7447 cdf35c40-ae34-48e0-9cc9-0c9da1808c22