Add a fourth parameter to the BP_Component::start() method.
This parameter is an array that allows extended classes to define some
properties that need to be set early one for BP_Component to be able
to work its magic.

Currently accepts 'adminbar_myaccount_order', which sets a custom
position for the component menu generated under the WP Toolbar's
"My Account" menu.  If this value isn't set, we set the value to 90,
which will generate the menu before the Settings menu is added.

Fixes #3769.

git-svn-id: http://buddypress.svn.wordpress.org/trunk@7454 cdf35c40-ae34-48e0-9cc9-0c9da1808c22