Don't canonical redirect page_on_front if bp_current_action() is populated
When looking at a BP component and bp_current_action() is empty, it means that
the item is (1) not a single item (like a group or member), and (2) not a feed
request. Therefore, it's a good way to infer whether a page is a directory
page, and ought to be considered for redirecting to directory page_on_front
settings.

This fixes the problem of /activity/feed/ being redirected to the front page
when the Activity directory was set to page_on_front.

Fixes #5211

Props r-a-y

git-svn-id: http://buddypress.svn.wordpress.org/trunk@7471 cdf35c40-ae34-48e0-9cc9-0c9da1808c22