Make sure activity comments can be viewed on single activity pages.
r7482 removed the activity comment count check, which led to activity
comments not being displayed on single activity pages when a user was
not logged in.

This commit adds a check to see if we're on a single activity page in the
activity entry template to allow activity comments to be shown.

See #7482.

git-svn-id: http://buddypress.svn.wordpress.org/trunk@7483 cdf35c40-ae34-48e0-9cc9-0c9da1808c22