Fake some critical globals when running unit tests that touch wp_mail()
The wp_mail() function requires access to $_SERVER['SERVER_NAME'], which is
not available when running the unit tests. We follow the WordPress core test
suite, by faking this value during automated testing.

Note that the long-term solution to this issue should be to break notification
functionality out of the wrapper functions that are used to request friendship,
etc. The current fix is something of an ugly workaround, just to get the tests
running properly when using E_STRICT.

git-svn-id: http://buddypress.svn.wordpress.org/trunk@7497 cdf35c40-ae34-48e0-9cc9-0c9da1808c22