Don't auto-format bp_get_total_member_count()
Filtering the output of bp_get_total_member_count() meant that the value was
double-encoded in some places, causing problems for certain locales and certain
numbers. Since number formatting is a display-related issue, we rely on the
template to do the number formatting, and leave the the output of
bp_get_total_member_count() in its raw form.

Fixes #5228

git-svn-id: http://buddypress.svn.wordpress.org/trunk@7516 cdf35c40-ae34-48e0-9cc9-0c9da1808c22