Cast args to a new variable in messages_notification_new_message()
r7621 made it possible for this function to accept a $message object in
addition to an array, by casting the arguments to an array. However, this
caused downstream problems when the $message object was passed by reference,
as in the case of 'messages_message_sent'. This changeset leaves the passed
object untouched.

Fixes #5273

git-svn-id: http://buddypress.svn.wordpress.org/trunk@7650 cdf35c40-ae34-48e0-9cc9-0c9da1808c22