Add delete_all support for all _delete_meta() functions
The delete_all parameter allows you to delete all meta items matching the
specified meta_key, regardless of the associated object.

Because our meta delete functions also support the deletion of all metadata
associated with an object (by leaving out the meta_key param), a decision had
to be made regarding the behavior when _delete_meta() is called with delete_all
= true and meta_key = false. The most logical (and least destructive) strategy
was deemed to be: force delete_all to false when meta_key is also false. This
ensures that you don't accidentally wipe out all metadata for a component.

See #5400

git-svn-id: http://buddypress.svn.wordpress.org/trunk@7883 cdf35c40-ae34-48e0-9cc9-0c9da1808c22