Don't test prev_value return value on WP < 3.7
There was a bug prior to WP 3.7 that caused update_metadata() to return true
even when the update failed due to a 'prev_value' parameter not being matched.
Since this is not really a functionality issue, but is more about error
reporting, it doesn't affect end-user functionality in BP, except insofar as
the function will return an unexpected value in those cases. Instead of coming
up with an arcane workaround for these old versions, we'll just skip the tests
for the time being.

See https://core.trac.wordpress.org/ticket/24933 for background on the WP issue

git-svn-id: http://buddypress.svn.wordpress.org/trunk@7887 cdf35c40-ae34-48e0-9cc9-0c9da1808c22