Make sure all xprofile fields are saved in wp-admin/profile
Unlike front-end edit-fields screens, the wp-admin/profile displays all groups of fields on a single page. So the list of field ids is an array gathering for each group of fields a distinct comma separated list of ids. As a result, before using the wp_parse_id_list() function, we must ensure that these ids are merged into a single comma separated list.

Fixes #5425

git-svn-id: http://buddypress.svn.wordpress.org/trunk@8028 cdf35c40-ae34-48e0-9cc9-0c9da1808c22