Prevent the last displayed activity on stream to be reloaded after a posted update
Since the introduction of the heartbeat feature to automatically check for new activities, when posting an update: an offset argument is used to load all new activities. This offset is given by the last activity id displayed on the stream, to be sure that specific activity is not reloaded, we need to increment this id by 1.

Fixes #5447

git-svn-id: http://buddypress.svn.wordpress.org/trunk@8048 cdf35c40-ae34-48e0-9cc9-0c9da1808c22