Make sure a deleted message thread is no more viewable by the user
When a user deletes a message thread, the message is no more shown in its inbox or outbox, but typing the url of the message thread will load the message giving the impression, it has not been deleted. Reason is a message is completely deleted as soon as all recipients and senders deleted it. So we need to make sure the user will be redirected to his inbox if he tries to view a message thread he deleted.

Props henrywright

Fixes #5254

git-svn-id: http://buddypress.svn.wordpress.org/trunk@8049 cdf35c40-ae34-48e0-9cc9-0c9da1808c22