BP_User_Query search_terms should match against user_login and user_nicename in addition to xprofile fields
This changeset changes the search_terms logic in BP_User_Query so that it
matches user_login and user_nicename in the core users table. Then, via filter,
the xprofile component amends the WHERE clause so that search_terms are matched
in the users table OR the xprofile data table.

Clauses have been refactored to subqueries rather than storing the matched
user IDs in PHP, for improved performance.

See #5155

git-svn-id: http://buddypress.svn.wordpress.org/trunk@8087 cdf35c40-ae34-48e0-9cc9-0c9da1808c22