Enable Spam admin tools on non-MS users.php Dashboard panel
This change brings non-MS installations of BuddyPress closer to the
functionality that is native to WP Multisite, with respect to the ability to
manage spam/ham users from the Users panel. It also addresses a shortcoming
whereby admins on non-MS sites would have no way at all to spam/unspam users
if the Settings component was disabled, since the front-end Spam tools are in
that optional component.

Fixes #5113

Props imath, r-a-y

git-svn-id: http://buddypress.svn.wordpress.org/trunk@8174 cdf35c40-ae34-48e0-9cc9-0c9da1808c22