Notifications: Invalidate cache when a notification is updated.
Object caching was added to the notifications component in r7813.
However, the cached item was not invalidated whenever a notification
is updated (either by using one of the marked as read functions or via
using the BP_Notifications_Notification::update() method).

This commit invalidates the cache directly in the update() class
method and adds a unit test.

See #5377

git-svn-id: http://buddypress.svn.wordpress.org/trunk@8176 cdf35c40-ae34-48e0-9cc9-0c9da1808c22